+++ app/src/main/java/jp/ac/it_college/std/ikemen/reachable/company/coupon/listener/OnActionClickListener.java
@@ -1,4 +1,4 @@
-package jp.ac.it_college.std.ikemen.reachable.company;
+package jp.ac.it_college.std.ikemen.reachable.company.coupon.listener;
 
 import android.view.View;
